JobScheduler

Description

Provide a queue (implemented as a TChan) and a configurable number of workers.

Use this module like so:

    do
        bracket (JobScheduler.init 4) (JobScheduler.shutdown 4) $ jobQueue ->
            -- code that uses the queue

This ensures the resources are cleaned up.

Documentation

type Task = Maybe JobSpec

A Task is just a Maybe-wrapped JobSpec. This allows us to switch on "something/nothing to do" in the workers. Receiving a Nothing as a Task means "you're done" to the worker.

type JobQueue = TChan Task

The Queue itself is a TChan of these Tasks. The semantics are what we would expect (and want). When a worker tries to read from an empty JobQueue, that worker will block until a value shows up on the TChan.

init :: Int -> IO JobQueue

Start up a worker pool of k members. Return the TChan that can be used to assign work to the pool.

shutdown :: Int -> JobQueue -> IO ()

Spam Nothing task into the job queue. When a worker receives a Nothing Task, it will quit.

enqueue :: MonadIO m => JobSpec -> JobQueue -> m ()

Add a job to the queue specified by a JobSpec.

reportWriter :: TChan String -> IO ()

Watch the report queue for messages and write them to the terminal

worker :: TChan String -> JobQueue -> Int -> IO ()

Run a worker. It can write to the report queue and it can read from the jobQueue. It is also assigned a sequential number as a label, so we know it's "worker 1," for example. The actual job is done in the performJob helper. This prints the job's name from the JobSpec and then waits the given duration in seconds. Then it prints that it's finished.

performJob :: JobSpec -> TChan String -> Int -> IO ()

Actually perform the specified job. In this case we just sleep for however long the job specifies.